upload image tentatively working when editing cards and with pngs
