+++ src/components/Modals/CardModal/CardModal.jsx
@@ -232,6 +232,7 @@ const CardModal = ({
                     images={images}
                     setImages={setImages}
                     isAdd
+                    cardId={cardId}
                   />
                 </>
               )}
