+++ src/screens/Library/LibraryPage.jsx
@@ -5,9 +5,6 @@ import StandardCardTable from "src/components/StandardCardTable";
 import useUser from "src/utils/lib/useUser";
 import PaginationTab from "../../components/PaginationTab";
 
-import { uploadFile, listBlobs } from "../../utils/blobStorage";
-import FilePicker from "chakra-ui-file-picker";
-
 const LibraryPage = ({ cardsFromDatabase, numPagesInitial }) => {
   const { user: currentUser } = useUser({
     redirectIfFound: false,
@@ -17,16 +14,7 @@ const LibraryPage = ({ cardsFromDatabase, numPagesInitial }) => {
   const [cards, setCards] = useState(cardsFromDatabase);
 
   const [searchString, setSearchString] = useState("");
-  const [fileList, setFileList] = useState([]);
 
-  const myRef = createRef();
-
-  const upload = () => {
-    const file = fileList[0];
-    uploadFile(file.name, file, { test_metadata: "test" }).then((res) => {
-      console.log(res._response.request.url);
-    });
-  };
   const [currentPage, setCurrentPage] = useState(1);
   const [numPages, setNumPages] = useState(numPagesInitial);
   const { searchedCards, handleSearch } = useSearch(
@@ -49,17 +37,6 @@ const LibraryPage = ({ cardsFromDatabase, numPagesInitial }) => {
         setCurrentPage={setCurrentPage}
       />
 
-      <Box>
-        <FilePicker
-          onFileChange={(fileList) => setFileList(fileList)}
-          placeholder="placeholder"
-          clearButtonLabel="Clear Selected Files"
-          multipleFiles={false}
-          hideClearButton={false}
-          ref={myRef}
-        />
-        <Button onClick={upload}>Upload</Button>
-      </Box>
       <StandardCardTable
         cards={cards}
         setCards={setCards}
